refactor: move asana.js path
