+++ src/hooks/asana/useProjects.js
@@ -1,4 +1,4 @@
-import { client } from '../asana.js'
+import { client } from './asana.js'
 import { useEffect, useState } from 'react'
 
 export function useProjects({ workspaceGid }) {
